Add files via upload
I was able to make the code on Google Colab, however I remember you guys saying I should use PyCharm. I made codes to both. The changes mainly lie in the file path and removing the pip install commands. I was not able to convert the PyCharm file into a notebook, every time that I did, it said the notebook was corrupted when I went to go upload it. Let me know if I need to make changes!